+++ lithops/constants.py
@@ -24,7 +24,7 @@ LOGGER_FORMAT = "%(asctime)s [%(levelname)s] %(name)s -- %(message)s"
 LOGGER_FORMAT_SHORT = "[%(levelname)s] %(filename)s -- %(message)s"
 LOGGER_LEVEL_CHOICES = ["debug", "info", "warning", "error", "critical"]
 
-STORAGE_CLI_MSG = '{} Storage client created'
+STORAGE_CLI_MSG = '{} client created'
 COMPUTE_CLI_MSG = '{} client created'
 
 LOCALHOST = 'localhost'